+++ src/contracts/entrypoints.rs
@@ -1,11 +1,12 @@
 use crate::{
-    contracts::{FeeTier, LiquidityTick, Pool, PoolKey, Position, PositionTick, Tick},
+    contracts::{
+        FeeTier, InvariantError, LiquidityTick, Pool, PoolKey, Position, PositionTick, Tick,
+    },
     invariant::{CalculateSwapResult, QuoteResult, SwapHop},
     math::{
         liquidity::Liquidity, percentage::Percentage, sqrt_price::SqrtPrice,
         token_amount::TokenAmount,
     },
-    InvariantError,
 };
 use alloc::vec::Vec;
 use ink::primitives::AccountId;
