+++ src/contracts/mod.rs
@@ -1,10 +1,12 @@
 pub mod collections;
 pub mod entrypoints;
+pub mod error;
 pub mod events;
 pub mod logic;
 pub mod storage;
 
 pub use collections::*;
 pub use entrypoints::*;
+pub use error::*;
 pub use events::*;
 pub use storage::*;
