+++ src/e2e/change_fee_receiver.rs
@@ -1,7 +1,7 @@
 #[cfg(test)]
 pub mod e2e_tests {
+    use crate::contracts::InvariantError;
     use crate::invariant::Invariant;
-    use crate::InvariantError;
     use crate::{
         contracts::{entrypoints::InvariantTrait, FeeTier, PoolKey},
         invariant::InvariantRef,
