+++ src/e2e/change_protocol_fee.rs
@@ -1,7 +1,7 @@
 #[cfg(test)]
 pub mod e2e_tests {
+    use crate::contracts::InvariantError;
     use crate::invariant::Invariant;
-    use crate::InvariantError;
     use crate::{
         contracts::entrypoints::InvariantTrait, invariant::InvariantRef,
         math::types::percentage::Percentage,
