+++ src/e2e/create_pool.rs
@@ -2,11 +2,10 @@
 pub mod e2e_tests {
     use crate::invariant::Invariant;
     use crate::{
-        contracts::{entrypoints::InvariantTrait, FeeTier},
+        contracts::{entrypoints::InvariantTrait, FeeTier, InvariantError},
         invariant::InvariantRef,
         math::types::percentage::Percentage,
         math::types::sqrt_price::{calculate_sqrt_price, SqrtPrice},
-        InvariantError,
     };
     use decimal::*;
     use ink_e2e::ContractsBackend;
