+++ src/e2e/interaction_with_pool_on_removed_fee_tier.rs
@@ -2,7 +2,7 @@
 pub mod e2e_tests {
     use crate::invariant::Invariant;
     use crate::{
-        contracts::{entrypoints::InvariantTrait, FeeTier, PoolKey},
+        contracts::{entrypoints::InvariantTrait, FeeTier, InvariantError, PoolKey},
         invariant::InvariantRef,
         math::{
             types::{
@@ -14,7 +14,6 @@ pub mod e2e_tests {
             },
             MIN_SQRT_PRICE,
         },
-        InvariantError,
     };
     use decimal::*;
     use ink::primitives::AccountId;
