+++ src/e2e/position_slippage.rs
@@ -2,14 +2,13 @@
 pub mod e2e_tests {
     use crate::invariant::Invariant;
     use crate::{
-        contracts::{entrypoints::InvariantTrait, FeeTier, PoolKey},
+        contracts::{entrypoints::InvariantTrait, FeeTier, InvariantError, PoolKey},
         invariant::InvariantRef,
         math::types::{
             liquidity::Liquidity,
             percentage::Percentage,
             sqrt_price::{calculate_sqrt_price, SqrtPrice},
         },
-        InvariantError,
     };
     use decimal::*;
     use ink_e2e::ContractsBackend;
