+++ src/lib.rs
@@ -5,38 +5,6 @@ mod contracts;
 pub mod e2e;
 pub mod math;
 
-#[derive(Debug, PartialEq, Eq)]
-#[ink::scale_derive(Encode, Decode, TypeInfo)]
-pub enum InvariantError {
-    NotAdmin,
-    NotFeeReceiver,
-    PoolAlreadyExist,
-    PoolNotFound,
-    TickAlreadyExist,
-    InvalidTickIndexOrTickSpacing,
-    PositionNotFound,
-    TickNotFound,
-    FeeTierNotFound,
-    PoolKeyNotFound,
-    AmountIsZero,
-    WrongLimit,
-    PriceLimitReached,
-    NoGainSwap,
-    InvalidTickSpacing,
-    FeeTierAlreadyExist,
-    PoolKeyAlreadyExist,
-    UnauthorizedFeeReceiver,
-    ZeroLiquidity,
-    TransferError,
-    TokensAreSame,
-    AmountUnderMinimumAmountOut,
-    InvalidFee,
-    NotEmptyTickDeinitialization,
-    InvalidInitTick,
-    InvalidInitSqrtPrice,
-    InvalidSize,
-    InvalidArithmeticOperation,
-}
 #[ink::contract]
 pub mod invariant {
     use crate::contracts::{
@@ -55,8 +23,8 @@ pub mod invariant {
     use crate::math::token_amount::TokenAmount;
     use crate::math::types::liquidity::Liquidity;
 
+    use crate::contracts::InvariantError;
     use crate::math::{compute_swap_step, MAX_SQRT_PRICE, MIN_SQRT_PRICE};
-    use crate::InvariantError;
     use decimal::*;
     use ink::contract_ref;
     use ink::prelude::vec;
